Split rendering into hpp and cpp
